Prevent successive physics frames from overlapping.
Not sure if this is actually a problem, but there's a lot going on in the car
physics, and it's possible that under some circumstances (on older hardware?)
it would still be chugging through the code by the time the next frame starts.

If that ever happens, then drop frames until the code from the previous one is
finished.