Fix water particles still spraying from under (invisible) wheels when a sunk car is being moved back to the track.
This is a minor glitch that was introduced after move to _physics_process
in commit 82517e45561de4d913a1e1991376d3c6895cae2e.
This caused the code for swtiching off particles to be skipped when
car is in "freeze" mode, so need to explicitly switch the particles off
when sinking.